pyproject.toml
[project]
name = "bareduckdb"
dynamic = ["version"]
description = "Minimal Cython-based DuckDB Bindings"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "typing-extensions>=4.15.0",
]
authors = [{name = "Paul T", email = "paul@iqmo.com"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Cython",
    "Programming Language :: Python :: Free Threading",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
arrow = ["pyarrow==22.0"]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.8.0",
    "pytest-cov>=6.0.0",
    "pytest-timeout>=2.3.1",
    "pytest-randomly>=4.0.1",
    "pytest-run-parallel>=0.2.0",
    "pytest-forked>=1.6.0",
    "pyarrow>=22.0",
    "pandas;python_version < '3.14'",  # For 3.14t manually install nightly pandas until 3.0 ships or PEP780 lands
]

[tool.pyright]
typeCheckingMode = "basic"
reportGeneralTypeIssues = false
reportPrivateImportUsage = false
reportAttributeAccessIssue = false
reportMissingTypeStubs = false
exclude = ["benchmark/**", "setup.py"]
venvPath = "."
venv = ".venv"
ignore = ["tests/**"]

[build-system]
requires = [
    "setuptools>=42",
    "setuptools-scm>=8",
    "wheel",
    "Cython>=3.0",
    "pyarrow>=22.0",  # Required at build time
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bareduckdb", "bareduckdb.core", "bareduckdb.core.impl", "bareduckdb.aio", "bareduckdb.compat",  "bareduckdb._libs", "bareduckdb.functional", "bareduckdb.dataset", "bareduckdb.dataset.impl"]

[tool.setuptools.package-data]
bareduckdb = ["_libs/libduckdb.*"]

[tool.setuptools_scm]
version_file = "bareduckdb/_version.py"

[tool.cibuildwheel]
build-frontend = "build[uv]"
manylinux-x86_64-image = "manylinux_2_28"
manylinux-aarch64-image = "manylinux_2_28"

[tool.cython-lint]
max-line-length = 160

[tool.ruff]
line-length = 160
exclude = [
  ".git",
  "__pycache__",
  "build",
  "dist",
  "docs/source/conf.py",
  "tests",
  "benchmark",
  "tests.ipynb",
  "analysis",
  "working",
  "benchmarks",
  "dev_*.py",
  "test_*.py",
  "*_repro.py",
  "setup.py",
  "notebooks",
  "external"
]
extend-include = ["*.ipynb"]

[tool.ruff.lint]
# pyflakes, pycodestyle, isort
select = ["F", "E", "W", "S", "B", "G", "N", "I001", "T", "PD", "C90"]
ignore = [
  # E203,
  # "W503",
  # "W293",
  "W291",
  "E501",
  "S608",
  "C901"
]

[tool.uv]
default-groups = ["dev"]
environments = [ # no need to resolve packages beyond these platforms with uv...
    "python_version >= '3.12' and sys_platform == 'darwin' and platform_machine == 'arm64'",
    "python_version >= '3.12' and sys_platform == 'linux' and platform_machine == 'x86_64'",
]
#required-environments = [ 
#    "python_version >= '3.13' and sys_platform == 'darwin' and platform_machine == 'arm64'",
#    "python_version >= '3.13' and sys_platform == 'linux' and platform_machine == 'x86_64'",
#]

prerelease = "if-necessary-or-explicit"

[dependency-groups]
dev = [
    "coverage>=7.11.0",
    "ipykernel>=7.0.1;python_version < '3.14'",
    #"magic-duckdb",  # Install with: uv pip install magic-duckdb --no-deps
    "pandas",
    "pandas-stubs",
    "polars>=1.34.0;python_version < '3.14'",
    "pre-commit-uv>=4.2.0",
    "pyarrow>=22.0",
    "pyarrow-stubs", # out of date, but good enough
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-forked>=1.6.0",
    "pytest-randomly>=4.0.1",
    "pytest-run-parallel>=0.2.0",
    "pytest-timeout>=2.3.1",
    "pytest-xdist>=3.8.0",
    "vortex-data;python_version < '3.14'", # no free-threaded
    "tabulate",
]
downstream = [
    "duckdb-engine>=0.17.0",
    "ibis-framework[duckdb]>=11.0.0",
    "ipython>=9.7.0",
    "jupysql>=0.11.1",
    "magic-duckdb>=0.2.3",
    "sqlalchemy>=2.0.44",
]

[tool.uv.extra-build-dependencies]
bareduckdb = ["Cython", "pyarrow>=22.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 90
# Ignore downstream tests by default (they require special installation)
addopts = [
    "-v",
    "--durations=5",
    "-n", "auto",
    "--maxprocesses=8",  # Cap workers to avoid MemoryError on large data tests
    "--maxfail=10",
    "--cov=bareduckdb",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--ignore=tests/downstream",
    "--ignore=tests/experimental",
    "--ignore=tests/comparison",

]
# For free-threaded Python: Convert GIL re-enablement warnings to errors
# This prevents INTERNALERROR in xdist and gives clear error messages
filterwarnings = [
    "error::RuntimeWarning",
    "ignore:invalid escape sequence:SyntaxWarning:sql.parse", # jupysql
]

[tool.coverage.run]
source = ["bareduckdb"]
branch = true
omit = [
    "*/tests/*",
    "*/benchmarks/*",
    "*/tests/downstream/*",
    "*.pyx",  # Cython source files
    "*/impl/*.pyx",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
skip_covered = false
skip_empty = true
precision = 2
